AI Image: Extract AI Image usage to hook
