+++ projects/plugins/jetpack/extensions/plugins/ai-assistant-plugin/components/ai-image/types.ts
@@ -1,0 +1,1 @@
+export const FEATURED_IMAGE_FEATURE_NAME = 'featured-post-image' as const;
